+++ backend/src/main/java/com/example/demo/services/socialevent/dto/EventQualificationDTO.java
@@ -1,4 +1,6 @@
 package com.example.demo.services.socialevent.dto;
 
-public record EventQualificationDTO(Long id, Double qualification) {
+import java.math.BigDecimal;
+
+public record EventQualificationDTO(Long id, BigDecimal qualification) {
 }
